refactor: 클래스 디렉토리 구분
