+++ app/src/main/java/dodeunifront/dodeuni/community/CommunityFragment.java
@@ -1,4 +1,4 @@
-package dodeunifront.dodeuni;
+package dodeunifront.dodeuni.community;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import dodeunifront.dodeuni.R;
 
 public class CommunityFragment extends Fragment {
